Fixed bugs

+++ app/src/main/java/com/paraskcd/unitedwalls/model/Uploader.kt
@@ -3,7 +3,7 @@ package com.paraskcd.unitedwalls.model
 data class Uploader(
     val __v: Int,
     val _id: String,
-    val userID: Int,
+    val userID: Long,
     val username: String,
     val avatar_file_url: String?,
     val avatar_uuid: String?,
